Changes toward the 1.30.7 version. See CHANGELOG for more details.

+++ R/featureCounts.R
@@ -1,29 +1,35 @@
-featureCounts <- function(files,annot.inbuilt="mm10",annot.ext=NULL,isGTFAnnotationFile=FALSE,GTF.featureType="exon",GTF.attrType="gene_id",chrAliases=NULL,useMetaFeatures=TRUE,allowMultiOverlap=FALSE,minOverlap=1,fracOverlap=0,fracOverlapFeature=0,largestOverlap=FALSE,nonOverlap=NULL,nonOverlapFeature=NULL,readExtension5=0,readExtension3=0,read2pos=NULL,countMultiMappingReads=FALSE,fraction=FALSE,isLongRead=FALSE,minMQS=0,splitOnly=FALSE,nonSplitOnly=FALSE,primaryOnly=FALSE,ignoreDup=FALSE,strandSpecific=0,juncCounts=FALSE,genome=NULL,isPairedEnd=FALSE,requireBothEndsMapped=FALSE,checkFragLength=FALSE,minFragLength=50,maxFragLength=600,countChimericFragments=TRUE,autosort=TRUE,nthreads=1,byReadGroup=FALSE,reportReads=NULL,reportReadsPath=NULL,maxMOp=10,tmpDir=".",verbose=FALSE)
+featureCounts <- function(files,annot.inbuilt="mm10",annot.ext=NULL,isGTFAnnotationFile=FALSE,GTF.featureType="exon",GTF.attrType="gene_id",GTF.attrType.extra=NULL,chrAliases=NULL,useMetaFeatures=TRUE,allowMultiOverlap=FALSE,minOverlap=1,fracOverlap=0,fracOverlapFeature=0,largestOverlap=FALSE,nonOverlap=NULL,nonOverlapFeature=NULL,readExtension5=0,readExtension3=0,read2pos=NULL,countMultiMappingReads=TRUE,fraction=FALSE,isLongRead=FALSE,minMQS=0,splitOnly=FALSE,nonSplitOnly=FALSE,primaryOnly=FALSE,ignoreDup=FALSE,strandSpecific=0,juncCounts=FALSE,genome=NULL,isPairedEnd=FALSE,requireBothEndsMapped=FALSE,checkFragLength=FALSE,minFragLength=50,maxFragLength=600,countChimericFragments=TRUE,autosort=TRUE,nthreads=1,byReadGroup=FALSE,reportReads=NULL,reportReadsPath=NULL,maxMOp=10,tmpDir=".",verbose=FALSE)
 {
 	flag <- FALSE
 	files <- normalizePath(files, mustWork=T)
 	if(!is.null(annot.ext) && is.character(annot.ext)) annot.ext <- normalizePath(annot.ext, mustWork=T)
 	if(!is.null(chrAliases))chrAliases <- normalizePath(chrAliases, mustWork=T)
 	if(!is.null(genome)) genome <- normalizePath(genome, mustWork=T)
-	if(!is.null(reportReadsPath))reportReadsPath <- normalizePath(reportReadsPath, mustWork=T)
+	if(!is.null(reportReadsPath)){
+		reportReadsPath <- normalizePath(reportReadsPath, mustWork=T)
+	}else reportReadsPath <- ' '
+	strandSpecific<-as.character(strandSpecific)
+	strandSpecific<-paste(strandSpecific, collapse=".")
+	strandSpecific<-gsub(",", ".", strandSpecific)
 
+	annot.screen.output <- 'R data.frame'
 	if(is.null(annot.ext)){
 	  switch(tolower(as.character(annot.inbuilt)),
 	    mm9={
 	      ann <- system.file("annot","mm9_RefSeq_exon.txt",package="Rsubread")
-	      cat("NCBI RefSeq annotation for mm9 (build 37.2) is used.\n")
+		  annot.screen.output <- 'inbuilt (mm9)'
 		},
 	    mm10={
 	      ann <- system.file("annot","mm10_RefSeq_exon.txt",package="Rsubread")
-	      cat("NCBI RefSeq annotation for mm10 (build 38.1) is used.\n")
+		  annot.screen.output <- 'inbuilt (mm10)'
 		 },
 	    hg19={
 	      ann <- system.file("annot","hg19_RefSeq_exon.txt",package="Rsubread")
-	      cat("NCBI RefSeq annotation for hg19 (build 37.2) is used.\n")
+		  annot.screen.output <- 'inbuilt (hg19)'
 	       },
 	    hg38={
 	      ann <- system.file("annot","hg38_RefSeq_exon.txt",package="Rsubread")
-	      cat("NCBI RefSeq annotation for hg38 (build 38.2) is used.\n")
+		  annot.screen.output <- 'inbuilt (hg38)'
 	       },
 	       {
 		stop("In-built annotation for ", annot.inbuilt, " is not available.\n")
@@ -33,6 +39,7 @@ featureCounts <- function(files,annot.inbuilt="mm10",annot.ext=NULL,isGTFAnnotat
 	else{
 	  if(is.character(annot.ext)){
 	    ann <- annot.ext
+		annot.screen.output <- paste0(basename(ann), " (", ifelse(isGTFAnnotationFile, "GTF", "SAF"), ")");
 	  }
 	  else{
 	    annot_df <- as.data.frame(annot.ext,stringsAsFactors=FALSE)
@@ -87,43 +94,50 @@ featureCounts <- function(files,annot.inbuilt="mm10",annot.ext=NULL,isGTFAnnotat
 	do_detection_calls <- FALSE
 	max_missing_bases_in_read <- -1
 	max_missing_bases_in_feature <- -1
+	GTF.attrType.extra_str <- " "
 	if(!is.null(nonOverlap)) max_missing_bases_in_read <- nonOverlap
 	if(!is.null(nonOverlapFeature)) max_missing_bases_in_feature <- nonOverlapFeature
+	if(!is.null(GTF.attrType.extra))GTF.attrType.extra_str <- paste(GTF.attrType.extra, collapse="\t")
 	  
-	cmd <- paste("readSummary",ann,files_C,fout,as.numeric(isPairedEnd),minFragLength,maxFragLength,0,as.numeric(allowMultiOverlap),as.numeric(useMetaFeatures),nthreads,as.numeric(isGTFAnnotationFile),strandSpecific,reportReads_C,as.numeric(requireBothEndsMapped),as.numeric(!countChimericFragments),as.numeric(checkFragLength),GTF.featureType,GTF.attrType,minMQS,as.numeric(countMultiMappingReads),chrAliases_C," ",as.numeric(FALSE),14,readExtension5,readExtension3,minOverlap,split_C,read2pos_C," ",as.numeric(ignoreDup),as.numeric(!autosort),as.numeric(fraction),as.numeric(largestOverlap),PE_orientation,as.numeric(juncCounts),genome_C,maxMOp,0,as.numeric(fracOverlap),as.character(tmpDir),"0",as.numeric(byReadGroup),as.numeric(isLongRead),as.numeric(verbose),as.numeric(fracOverlapFeature), as.numeric(do_detection_calls), as.numeric(max_missing_bases_in_read), as.numeric(max_missing_bases_in_feature), as.numeric(primaryOnly), reportReadsPath,sep=",")
+	cmd <- paste("readSummary",ann,files_C,fout,as.numeric(isPairedEnd),minFragLength,maxFragLength,0,as.numeric(allowMultiOverlap),as.numeric(useMetaFeatures),nthreads,as.numeric(isGTFAnnotationFile),strandSpecific,reportReads_C,as.numeric(requireBothEndsMapped),as.numeric(!countChimericFragments),as.numeric(checkFragLength),GTF.featureType,GTF.attrType,minMQS,as.numeric(countMultiMappingReads),chrAliases_C," ",as.numeric(FALSE),14,readExtension5,readExtension3,minOverlap,split_C,read2pos_C," ",as.numeric(ignoreDup),as.numeric(!autosort),as.numeric(fraction),as.numeric(largestOverlap),PE_orientation,as.numeric(juncCounts),genome_C,maxMOp,0,as.numeric(fracOverlap),as.character(tmpDir),"0",as.numeric(byReadGroup),as.numeric(isLongRead),as.numeric(verbose),as.numeric(fracOverlapFeature), as.numeric(do_detection_calls), as.numeric(max_missing_bases_in_read), as.numeric(max_missing_bases_in_feature), as.numeric(primaryOnly), reportReadsPath, GTF.attrType.extra_str, annot.screen.output ,sep=",")
 	n <- length(unlist(strsplit(cmd,",")))
 	C_args <- .C("R_readSummary_wrapper",as.integer(n),as.character(cmd),PACKAGE="Rsubread")
 
-	x <- read.delim(fout,stringsAsFactors=FALSE)
-	colnames(x)[1:6] <- c("GeneID","Chr","Start","End","Strand","Length")
+    if(file.exists(fout)){
+		x <- read.delim(fout,stringsAsFactors=FALSE)
+		colnames(x)[1:6] <- c("GeneID","Chr","Start","End","Strand","Length")
 
-	x_summary <- read.delim(paste(fout,".summary",sep=""), stringsAsFactors=FALSE)
+		x_summary <- read.delim(paste(fout,".summary",sep=""), stringsAsFactors=FALSE)
 
-	if(juncCounts)
-		x_jcounts <- read.delim(paste(fout,".jcounts",sep=""), stringsAsFactors=FALSE)
+		if(juncCounts)
+			x_jcounts <- read.delim(paste(fout,".jcounts",sep=""), stringsAsFactors=FALSE)
 
-	file.remove(fout)
-	file.remove(paste(fout,".summary",sep=""))
+		file.remove(fout)
+		file.remove(paste(fout,".summary",sep=""))
 
-	if(juncCounts)
-		file.remove(paste(fout,".jcounts",sep=""))
-	
-	if(flag) 
-	  file.remove(fout_annot)
-	
-	if(ncol(x) == 6){
-	  stop("No count data were generated.")
+		if(juncCounts)
+			file.remove(paste(fout,".jcounts",sep=""))
+		
+		if(flag) 
+		  file.remove(fout_annot)
+
+		add_attr_numb <- 0
+		if(!is.null(GTF.attrType.extra)) add_attr_numb <- length(GTF.attrType.extra)
+		if(ncol(x) <= (6 + add_attr_numb)){
+		  stop("No count data were generated.")
+		}
+		
+		y <- as.matrix(x[,-c(1:(6 + add_attr_numb))])
+		colnames(y) <- colnames(x)[-c(1:(6 + add_attr_numb))]
+		rownames(y) <- x$GeneID
+
+		if(juncCounts)
+			z <- list(counts=y,counts_junction=x_jcounts,annotation=x[,1:(6 + add_attr_numb)],targets=colnames(y),stat=x_summary)
+		else
+			z <- list(counts=y,annotation=x[,1:(6 + add_attr_numb)],targets=colnames(y),stat=x_summary)	
+		z	
+	}else{
+		stop("No counts were generated.")
 	}
-	
-	y <- as.matrix(x[,-c(1:6)])
-	colnames(y) <- colnames(x)[-c(1:6)]
-	rownames(y) <- x$GeneID
-
-	if(juncCounts)
-		z <- list(counts=y,counts_junction=x_jcounts,annotation=x[,1:6],targets=colnames(y),stat=x_summary)
-	else
-		z <- list(counts=y,annotation=x[,1:6],targets=colnames(y),stat=x_summary)	
-
-	z	
 }
